app_engine/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
grpcio>=1.11.0
slackclient>=1.2.1
slackeventsapi>=1.1.0
google-api-python-client>=1.7.4
google-cloud-storage>=1.10.0
google-cloud-datastore>=1.6.0
google-cloud-dlp>=0.7.0
google-cloud-bigquery>=1.5.0
